Introduce Zustand
- Add state store
- Add config load action
- Remove config context provider
- Remove old references to provider to use state store instead.
