+++ package.json
@@ -26,7 +26,8 @@
     "react-select": "^5.3.2",
     "react-table": "^7.8.0",
     "url-join": "^5.0.0",
-    "use-immer": "^0.7.0"
+    "use-immer": "^0.7.0",
+    "zustand": "^4.4.7"
   },
   "devDependencies": {
     "husky": "^8.0.3",
