+++ src/components/main/App/App.js
@@ -1,19 +1,27 @@
-import React from "react";
-import { Container } from "react-bootstrap";
+import React, { useEffect } from 'react';
+import { Container } from 'react-bootstrap';
 
 import Disclaimer from "../../info/Disclaimer";
 import Header from "../Header/Header";
 import Body from "../Body";
 import useInitializeApp from "./app-initialization";
+import { useStore } from '../../../state/state-store';
 
-import "./App.css";
-import ConfigContext, { useFetchConfigContext } from "../ConfigContext";
+import './App.css';
+
 
 export default function App() {
   // must be invoked before any other items dependent on context.
-  const [config, configErrorMessage] = useFetchConfigContext({});
+  const initialize = useStore(state => state.initialize);
+  const isConfigLoaded = useStore(state => state.isConfigLoaded);
+  const configErrorMessage = useStore(state => state.configError);
+  const config = useStore(state => state.config);
 
-  useInitializeApp(config);
+  useEffect(() => {
+    if (!isConfigLoaded() && configErrorMessage === null) {
+      initialize();
+    }
+  });
 
   if (configErrorMessage !== null) {
     return <div>{configErrorMessage}</div>;
@@ -23,13 +31,13 @@ export default function App() {
     return <div>Loading configuration...</div>;
   }
 
+  useInitializeApp(config);
+
   return (
-    <ConfigContext.Provider value={config}>
-      <Container fluid className="App">
-        <Disclaimer />
-        <Header />
-        <Body />
-      </Container>
-    </ConfigContext.Provider>
+    <Container fluid className="App">
+      <Disclaimer />
+      <Header />
+      <Body />
+    </Container>
   );
 }
