+++ src/components/main/Body/Body.js
@@ -33,7 +33,7 @@ import StationFilters, {
   useStationFiltering,
 } from "../../controls/StationFilters";
 import baseMaps from "../../maps/baseMaps";
-import { useConfigContext } from "../ConfigContext";
+import { useStore } from "../../../state/state-store";
 
 logger.configure({ active: true });
 
@@ -42,7 +42,7 @@ logger.configure({ active: true });
 // of debug callbacks for controlling station fetches. This hook is used only
 // by component Body; it's a way of clarifying and simplifying its code.
 function useMetadata() {
-  const config = useConfigContext();
+  const config = useStore(state => state.config);
 
   // Fetched metadata
   const [metadata, setMetadata] = useImmerByKey({
@@ -103,7 +103,7 @@ function useMetadata() {
 }
 
 function Body() {
-  const config = useConfigContext();
+  const config = useStore(state => state.config);
 
   // Metadata fetched from backend
   const { metadata, setStnsLimit, reloadStations } = useMetadata();
