+++ src/components/main/Header/Header.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { Row, Col } from "react-bootstrap";
-import { useConfigContext } from "../ConfigContext";
+import { useStore } from "../../../state/state-store";
 import "./Header.css";
 
 function Header() {
-  const config = useConfigContext();
+  const config = useStore(state => state.config);
   return (
     <Row className={"Header"}>
       <Col lg={3} className="text-left">
