+++ src/state/state-store.js
@@ -1,0 +1,15 @@
+import { create } from 'zustand';
+import { loadConfigAction } from './action-load-config';
+
+
+export const useStore = create((set, get) => ({
+    // States
+    config: null,
+    configError: null,
+    isConfigLoaded: () => get().config !== null,
+    // Actions
+    initialize: () => get()._loadConfig(),
+
+    // Private Actions
+    _loadConfig: loadConfigAction(set, get)
+  }));
